+++ client/src/CSS/BentoBox.css
@@ -1,0 +1,169 @@
+/* Bento Box Container */
+.bento-box {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 1.5rem; 
+    padding: 1.5rem; 
+    max-width: 1440px; 
+    margin: 1.5rem auto; 
+    box-sizing: border-box;
+    border-radius: 1.25rem; 
+}
+
+
+.col-left {
+    flex: 2 2 450px; 
+    display: flex;
+    flex-direction: column;
+    gap: 1rem; 
+    border-radius: 1rem;
+    overflow: hidden;
+}
+
+/* Right Column */
+.col-right {
+    flex: 1 1 300px; 
+    display: flex;
+    flex-direction: column;
+    gap: 1rem; 
+    border-radius: 1rem; 
+    overflow: hidden;
+}
+
+/* Rows within Columns */
+.col-row.row1 {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    background-color: var(--clr-surface-a10); 
+    border-radius: 0.8rem;
+    box-shadow: 0 3px 8px var(--color-shadow);
+    transition: box-shadow 0.3s ease, transform 0.3s ease;
+    overflow: hidden;
+}
+
+.col-row.row2 {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    background-color: var(--clr-primary-a20);
+    border-radius: 0.8rem;
+    box-shadow: 0 3px 8px var(--color-shadow);
+    transition: box-shadow 0.3s ease, transform 0.3s ease;
+    overflow: hidden;
+}
+
+.col-row.top {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    background-color: var(--clr-primary-a40);
+    border-radius: 0.8rem;
+    box-shadow: 0 3px 8px var(--color-shadow);
+    transition: box-shadow 0.3s ease, transform 0.3s ease;
+    overflow: hidden;
+}
+
+.col-row.bottom {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    background-color: var(--clr-surface-a30); 
+    border-radius: 0.8rem;
+    box--shadow: 0 3px 8px var(--color-shadow);
+    transition: box-shadow 0.3s ease, transform 0.3s ease;
+    overflow: hidden;
+}
+
+.col-row:hover {
+    box-shadow: 0 6px 14px rgba(0, 0, 0, 0.08);
+    transform: translateY(-3px); 
+}
+
+/* Titles */
+.threadvisted h1,
+.threadcreated h1,
+.threadtrending h1,
+.credibility h1 {
+    font-size: 1.5rem; 
+    font-weight: 600;
+    color: #333;
+    padding: 1rem; 
+    border-bottom: 1px solid #eee;
+    margin: 0;
+    letter-spacing: -0.01em; 
+}
+
+/* Content Areas */
+.threadvisted-content,
+.threadcreated-content,
+.threadtrending-content,
+.cred-content {
+    padding: 1rem; 
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    gap: 0.6rem; 
+}
+
+/* Paragraphs in Content Areas */
+.threadvisted-content p,
+.threadcreated-content p,
+.threadtrending-content p {
+    background-color: #f0f0f0;
+    padding: 0.8rem; 
+    border-radius: 0.6rem; 
+    font-size: 0.9rem; 
+    color: #555;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    cursor: pointer;
+    transition: background-color 0.3s ease, transform 0.3s ease;
+    line-height: 1.5; 
+}
+
+.threadvisted-content p:hover,
+.threadcreated-content p:hover,
+.threadtrending-content p:hover {
+    background-color: #e0e0e0;
+    transform: scale(1.01); 
+}
+
+h1.fire {
+    font-size: 1.8rem; 
+    animation: flicker 2s infinite;
+}
+
+@keyframes flicker {
+    0%, 100% { text-shadow: 0 -1px 3px #fff, 0 -2px 6px #FF3, 0 -6px 12px #F90, 0 -10px 20px #C33; }
+    50% { text-shadow: 0 -2px 6px #fff, 0 -4px 12px #FF3, 0 -8px 16px #F90, 0 -12px 24px #C33; }
+}
+
+/* Responsive Adjustments */
+@media (max-width: 768px) {
+    .bento-box {
+        flex-direction: column;
+        padding: 1.5rem;
+        margin: 1rem auto;
+    }
+
+    .col-left,
+    .col-right {
+        width: 100%;
+        min-width: fit-content;
+    }
+
+    .threadvisted h1,
+    .threadcreated h1,
+    .threadtrending h1,
+    .credibility h1 {
+        font-size: 1.5rem;
+    }
+
+    .threadvisted-content p,
+    .threadcreated-content p,
+    .threadtrending-content p {
+        font-size: 0.9rem;
+    }
+}
