+++ client/src/Manager.js
